+++ volo/src/lib.rs
@@ -4,9 +4,6 @@
 #![cfg_attr(not(doctest), doc = include_str!("../README.md"))]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
-#[macro_use]
-mod cfg;
-
 pub use motore::{layer, layer::Layer, service, Service};
 pub use tokio::main;
